Avoid setting global serializer on 'before' hook
We ran into an issue where having a mix of both REST and non-REST
endpoints would cause the non-REST endpoints to use the last used
serializer due to the setting being leaked/global from the last
request. While this isn't an issue for some applications (eg. pure
APIs), it does create issues when mixing.

Since we're recreating a response object on the hook, we can as
well just use the dancer2 core factory to manually create the
serializer (so it uses the global config), and pass it inside the
new response.